blob_detection/requirements.txt
matplotlib==3.9.2
numpy==2.1.2
Pillow==10.4.0
scipy==1.14.1
torch==2.4.0
torchvision==0.19.0
